init azp feature
