+++ Algorithms/azp.cpp
@@ -1,0 +1,15 @@
+#include <vector>
+
+#include "azp.h"
+
+template <typename T>
+ZoneControl<T>::ZoneControl()
+{
+
+}
+
+template <typename T>
+ZoneControl<T>::~ZoneControl()
+{
+
+}
