+++ Algorithms/pam.h
@@ -15,74 +15,12 @@
 
 #include <vector>
 #include <boost/unordered_map.hpp>
+
+#include "DataUtils.h"
+#include "rng.h"
+
 using namespace boost;
 
-/**
- * title = "xoroshiro+ / xorshift* / xorshift+ generators and the PRNG shootout", //
- * booktitle = "Online", //
- * url = "http://xoroshiro.di.unimi.it/", //
- */
-class Xoroshiro128Random
-{
-    long long s0;
-    long long s1;
-public:
-    Xoroshiro128Random(long long xor64 = 123456789) {
-        // set seed
-        // XorShift64* generator to seed:
-        if (xor64 == 0)
-            xor64 = 4101842887655102017L;
-        xor64 ^= (unsigned long long)xor64 >> 12; // a
-        xor64 ^= xor64 << 25; // b
-        xor64 ^= (unsigned long long)xor64 >> 27; // c
-        s0 = xor64 * 2685821657736338717L;
-        xor64 ^= (unsigned long long)xor64 >> 12; // a
-        xor64 ^= xor64 << 25; // b
-        xor64 ^= (unsigned long long)xor64 >> 27; // c
-        s1 = xor64 * 2685821657736338717L;
-    }
-    virtual ~Xoroshiro128Random() {}
-    int nextInt(int n) {
-        if (n <=0) return 0;
-        int r =  (int)((n & -n) == n ? nextLong() & n - 1 // power of two
-            : (unsigned long long)(((unsigned long long)nextLong() >> 32) * n) >> 32);
-        return r;
-    }
-    long long nextLong() {
-        long long t0 = s0, t1 = s1;
-        long long result = t0 + t1;
-        t1 ^= t0;
-        // left rotate: (n << d)|(n >> (INT_BITS - d));
-        s0 = (t0 << 55) | ((unsigned long long)t0 >> (64 - 55));
-        s0 = s0 ^ t1 ^ (t1 << 14); // a, b
-        s1 = (t1 << 36) | ((unsigned long long)t1 >> (64 -36));
-        return result;
-    }
-    double nextDouble() {
-#ifdef WIN32
-		char tempStr[] = "0x1.0p-53";
-		double nd = std::strtod(tempStr, NULL);
-        return ((unsigned long long)nextLong() >> 11) * nd;
-#else
-		return ((unsigned long long)nextLong() >> 11) * 0x1.0p-53;
-#endif
-    }
-    std::vector<int> randomSample(int samplesize, int n)
-    {
-        std::vector<int> samples(samplesize);
-        int i=0;
-        boost::unordered_map<int, bool> sample_dict;
-        boost::unordered_map<int, bool>::iterator it;
-        while (sample_dict.size() < samplesize) {
-            int rnd = nextInt(n);
-            if (sample_dict.find(rnd) == sample_dict.end()) {
-                samples[i++] = rnd;
-            }
-            sample_dict[rnd] = true;
-        }
-        return samples;
-    }
-};
 
 class PAMUtils
 {
@@ -90,72 +28,6 @@ public:
     static std::vector<int> randomSample(Xoroshiro128Random& rand,
                                          int samplesize, int n,
                                          const std::vector<int>& previous = std::vector<int>());
-};
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-///
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-class DistMatrix
-{
-protected:
-    std::vector<int> ids;
-    bool has_ids;
-public:
-    DistMatrix(const std::vector<int>& _ids=std::vector<int>())
-    : ids(_ids), has_ids(!_ids.empty()) {}
-    virtual ~DistMatrix() {}
-    // Get distance between i-th and j-th object
-    // if ids vector is provided, the distance (i,j) -> distance(ids[i], ids[j])
-    virtual double getDistance(int i, int j) = 0;
-    virtual void setIds(const std::vector<int>& _ids) {
-        ids = _ids;
-        has_ids = !ids.empty();
-    }
-};
-
-class RawDistMatrix : public DistMatrix
-{
-    double** dist;
-public:
-    RawDistMatrix(double** dist, const std::vector<int>& _ids=std::vector<int>())
-    : DistMatrix(_ids), dist(dist) {}
-    virtual ~RawDistMatrix() {}
-    virtual double getDistance(int i, int j) {
-        if (i == j) return 0;
-        if (has_ids) {
-            i = ids[i];
-            j = ids[j];
-        }
-        // lower part triangle
-        int r = i > j ? i : j;
-        int c = i < j ? i : j;
-        return dist[r][c];
-    }
-};
-
-class RDistMatrix : public DistMatrix
-{
-    int num_obs;
-    int n;
-    const std::vector<double>& dist;
-public:
-    RDistMatrix(int num_obs, const std::vector<double>& dist, const std::vector<int>& _ids=std::vector<int>())
-    : DistMatrix(_ids), num_obs(num_obs), dist(dist) {
-        n = (num_obs - 1) * num_obs / 2;
-    }
-    virtual ~RDistMatrix() {}
-    
-    virtual double getDistance(int i, int j) {
-        if (i == j) return 0;
-        if (has_ids) {
-            i = ids[i];
-            j = ids[j];
-        }
-        // lower part triangle, store column wise
-        int r = i > j ? i : j;
-        int c = i < j ? i : j;
-        int idx = n - (num_obs - c - 1) * (num_obs - c) / 2 + (r -c) -1 ;
-        return dist[idx];
-    }
 };
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
